(mu) make mu.h
